+++ RogueFrontier/Types/SystemType.cs
@@ -413,7 +413,7 @@ public record LightGenerator() : IGridGenerator<Color> {
         this.lc = lc;
         this.radius = radius;
     }
-    public Color Generate((int, int) p) {
+    public Color Generate((long, long) p) {
         //var xy = new XY(p);
         return new Color(255, 255, 204, Math.Min(255, (int)(radius * 255 / ((lc.pos - p).magnitude + 1))));
     }
